+++ class.tx_rnbase_configurations.php
@@ -73,6 +73,7 @@
  * @subpackage tx_rnbase
  *
  */
+tx_rnbase::load('tx_rnbase_util_Network');
 
 class tx_rnbase_configurations {
   // We store all Data to an internal ArrayObject
@@ -360,7 +361,7 @@ class tx_rnbase_configurations {
   function &getFlexFormArray() {
     static $flex;
     if (!is_array($flex)) {
-      $flex = t3lib_div::getURL(tx_rnbase_util_Extensions::extPath($this->getExtensionKey()) . $this->get('flexform'));
+      $flex = tx_rnbase_util_Network::getURL(tx_rnbase_util_Extensions::extPath($this->getExtensionKey()) . $this->get('flexform'));
       $flex = t3lib_div::xml2array($flex);
     }
     return $flex;