+++ mod/class.tx_rnbase_mod_BaseModFunc.php
@@ -22,7 +22,7 @@
 *
 *  This copyright notice MUST APPEAR in all copies of the script!
 ***************************************************************/
-
+tx_rnbase::load('tx_rnbase_util_Network');
 tx_rnbase::load('tx_rnbase_mod_IModule');
 tx_rnbase::load('tx_rnbase_mod_IModFunc');
 tx_rnbase::load('tx_rnbase_util_BaseMarker');
@@ -48,7 +48,7 @@ abstract class tx_rnbase_mod_BaseModFunc implements tx_rnbase_mod_IModFunc {
 		$conf = $this->getModule()->getConfigurations();
 
 		$file = t3lib_div::getFileAbsFileName($conf->get($this->getConfId().'template'));
-		$templateCode = t3lib_div::getURL($file);
+		$templateCode = tx_rnbase_util_Network::getURL($file);
 		if(!$templateCode) return $conf->getLL('msg_template_not_found').'<br />File: \'' . $file . '\'<br />ConfId: \'' . $this->getConfId().'template\'';
 		$subpart = '###'.strtoupper($this->getFuncId()).'###';
 		$template = tx_rnbase_util_Templates::getSubpart($templateCode, $subpart);