+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.15.19 (??.??.????)
  * new Wrapper Class Tx_Rnbase_Service_Base for \TYPO3\CMS\Core\Service\AbstractService or t3lib_svbase which is compatible for all TYPO3 versions
  * new Wrapper Class Tx_Rnbase_Backend_Utility for \TYPO3\CMS\Backend\Utility\BackendUtility or t3lib_BEfunc which is compatible for all TYPO3 versions
  * new Wrapper function tx_rnbase_util_Strings::inList
+ * new Wrapper function tx_rnbase_util_Network::getUrl
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall